requirements.txt
requests>=2.31.0
numpy>=1.26.0
pillow>=10.0.0
